Move ASCII util functions from lib/util.c to lib/ascii.c
