+++ gumbo-parser/src/svg_attrs.gperf
@@ -1,7 +1,7 @@
 %{
 #include "replacement.h"
 #include "macros.h"
-#include "util.h"
+#include "ascii.h"
 %}
 
 %ignore-case
