+++ gumbo-parser/src/svg_tags.gperf
@@ -1,7 +1,7 @@
 %{
 #include "replacement.h"
 #include "macros.h"
-#include "util.h"
+#include "ascii.h"
 %}
 
 %ignore-case
