+++ gumbo-parser/src/tag_lookup.c
@@ -5,7 +5,7 @@
 
 #include "tag_lookup.h"
 #include "macros.h"
-#include "util.h"
+#include "ascii.h"
 #include <string.h>
 
 #define TOTAL_KEYWORDS 150
