+++ gumbo-parser/src/tag_lookup.gperf
@@ -1,7 +1,7 @@
 %{
 #include "tag_lookup.h"
 #include "macros.h"
-#include "util.h"
+#include "ascii.h"
 %}
 
 %ignore-case
